Started moving stuff into the schemes
Looks nice, but now the tests broke due to the relative import
